Fixed macOS and Linux build for embedded libzim binary
Given using embedded libzim binary is the main use-case, I have disabled the use
of the system-installed one (which had precedence).
This can be changed manually at `libzim_local` in `binding.gyp`. We could use an option
or an environment variable to toggle it later but that seems secondary.

Linux fixes:
- rpath is correctly set (to `$ORIGIN`) so that it looks for libzim next to binding.
- libraries path updated to remove need for build-time symlinks.
- removed symlink to libzim bin in download step (launched right after an async call – not ideal).
- libzim binary is copied post-build next to the output file.,

macOS fixes:
- using same download folder with proper path.
- adding rpath to linking (not suffiscient).
- removed copy of libzim bin to system path.
- libzim binary is copied post-build next to the output file.
- output file's rpath to libzim is fixed to look for it aside.

General fixes:
- better variables usage in download step
- added a new post-build step (called bundle) to copy binaries and fix rpath